Set Up Primary Pages
I set up my homepage and jar ice cream page. I added to the readme page.

I put a link to the recipe jar ice cream on the homepage and set up the
jar ice cream page. I added a quick intro text to the homepage as well.

In the readme file, I wrote the purpose of the website and that I will add a
reflection to it later as well as stylize it with CSS later.